requirements.txt
alembic==1.9.0
cachetools==5.2.0
fastapi[all]==0.88.0
flask-swagger-ui==3.36.0
Flask[async]==2.2.2
json-cfg==0.4.2
marshmallow-sqlalchemy==0.28.1
marshmallow==3.19.0
openpyxl==3.0.9
Pillow==9.3.0
psycopg==3.1.7
psycopg2==2.9.5
pydantic==1.10.2
pyjwt==2.6.0
pymongo~=3.11.1
pytest-postgresql==4.1.1
python-dateutil==2.8.2
python-memcached==1.59
python-multipart==0.0.5
rdkit==2022.3.5
requests==2.28.1
SQLAlchemy==1.4.45
urllib3==1.26.9
uvicorn[standard]==0.20.0
yamlreader==3.0.4
